functions/job-trigger/requirements.txt
# Required for the Jobs API
google-cloud-run